// src/app/shared/models/JobcalanderIHQCRreport.model.ts
export class JobcalanderIHQCReport {
    id: number;
    stepno: number;
    jobname: string;
    qcdatetime: string;
    qcinspectorname: string;
    okqty: string;
    rejectedqty: string;
    segregationqty: string;
    reworkqty:  string;
       // instrument methods/name
    measurementtech: string;
    instrumentid : string;
    instrumentname:string;
    comments:string;
    status: string;
    createdby: string;
    updatedby: string;

}